Split mpi.hpp into multiple header files
